check function moved to homepage
